+++ unittests/cpp/gRPC/tests/LanguagePackRequest_test.cpp
@@ -94,7 +94,7 @@ class gRPCLanguagePackRequestTests : public GRPCTest {
     /**
      * Makes an async RPC and checks the requestStart/requestReceived values read by the handler.
      */
-    void testRPCTimestamps(std::string input, long expected) {
+    void testRPCTimestamps(std::string input, long long expected) {
         // Create a new client context for this call to avoid metadata from previous calls
         grpc::ClientContext context;
         context.AddMetadata("request-start", input);
