+++ unittests/cpp/gRPC/tests/MultiSetValue_test.cpp
@@ -99,7 +99,7 @@ class gRPCMultiSetValueTests : public GRPCTest {
     /**
      * Makes an async RPC and checks the requestStart/requestReceived values read by the handler.
      */
-    void testRPCTimestamps(std::string input, long expected) {
+    void testRPCTimestamps(std::string input, long long expected) {
         // Create a new client context for this call to avoid metadata from previous calls
         grpc::ClientContext context;
         context.AddMetadata("request-start", input);
